+++ src/main/java/pl/thedeem/intellij/dql/exec/runConfiguration/ExecuteDQLRunConfiguration.java
@@ -81,15 +81,16 @@ public class ExecuteDQLRunConfiguration extends RunConfigurationBase<ExecuteDQLR
         if (StringUtil.isEmpty(options.getSelectedTenant())) {
             throw new RuntimeConfigurationError(DQLBundle.message("runConfiguration.executeDQL.settings.errors.emptyTenant"));
         }
+        if (StringUtil.isEmpty(getName())) {
+            throw new RuntimeConfigurationError(DQLBundle.message("runConfiguration.executeDQL.settings.errors.emptyRunConfigName"));
+        }
         if (options.getFromFileSelected()) {
             String queryFromFile = getQueryFromFile(options.getDqlPath(), getProject());
             if (StringUtil.isEmpty(queryFromFile)) {
                 throw new RuntimeConfigurationError(DQLBundle.message("runConfiguration.executeDQL.settings.errors.emptyFile"));
             }
-        } else {
-            if (StringUtil.isEmpty(options.getDqlQuery())) {
-                throw new RuntimeConfigurationError(DQLBundle.message("runConfiguration.executeDQL.settings.errors.emptyQuery"));
-            }
+        } else if (StringUtil.isEmpty(options.getDqlQuery())) {
+            throw new RuntimeConfigurationError(DQLBundle.message("runConfiguration.executeDQL.settings.errors.emptyQuery"));
         }
     }