+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - DQL fragment selected during the query execution will now have a green border instead of a selection background
   - Adding data toolbar for DQL execution results opened as a new editor tab
   - Adding an option to change the DQL execution results column order and visibility
+  - Adding predefined timeframes to the DQL execution toolbar (user can still define a custom one)
 - The "Show DQL query" option when executing DQL will now try to show the parsed query instead of the raw one, if
   possible.
 - Adding support for "Expression DQL" file (with `.edql` extensions), which allows to define a DQL expression without